+++ dash/card-upload-act.php
@@ -26,7 +26,7 @@
         }
 
         // Apagando todas as sessões desse grupo
-        $sql = sprintf('delete from car_study_session where stud_id in (select stud_id from car_study where deck_id = %d and user_id = %d)', $deck_id, $user_id);
+        $sql = sprintf('delete from car_study_session where user_id = %d and stud_id in (select stud_id from car_study where deck_id = %d and user_id = %d)', $user_id, $deck_id, $user_id);
 
         $result = $mysqli->query($sql);
