+++ dash/deck-delete-act.php
@@ -29,12 +29,12 @@
 		}
 		
 		// Apagando todos as sessões de estudo
-		$sql = sprintf(' delete from car_study_session where stud_id in (select stud_id from car_study where deck_id = %d)', $deck_id);
+		$sql = sprintf(' delete from car_study_session where user_id = %d and stud_id in (select stud_id from car_study where deck_id = %d and user_id = %d)', $user_id, $deck_id, $user_id);
 		
 		$result = $mysqli->query($sql);
 		
 		// Apagando todos os estudos
-		$sql = sprintf(' delete from car_study where deck_id = %d', $deck_id);
+		$sql = sprintf(' delete from car_study where deck_id = %d and user_id = %d', $deck_id, $user_id);
 		
 		$result = $mysqli->query($sql);
