+++ dash/study-srs-new-act.php
@@ -56,7 +56,7 @@
 		
 		if ($user_count_study < $user_max_study) {
 			// Apaga todos os cards em branco do usuarios
-			$_sql = sprintf(" delete from car_card where deck_id = %d and (card_front is null or trim(card_front) = '' or card_back is null or trim(card_back) = '');", $deck_id);
+			$_sql = sprintf(" delete from car_card where deck_id = %d and user_id = %d and (card_front is null or trim(card_front) = '' or card_back is null or trim(card_back) = '');", $deck_id, $user_id);
 			
 			$_result = $mysqli->query($_sql);
 			
@@ -128,11 +128,13 @@
                                 select row_number() over (order by RAND()) as stse_order, %d, %d, card_id
                                   from car_card
                                  where deck_id = %d
+                                   and user_id = %d
                                    and (card_rate < %d or card_sequence < %d or card_last_study < DATE_SUB(NOW(), INTERVAL %d DAY))
                                  limit %d',
                                 $user_id,
                                 $stud_id,
                                 $deck_id,
+                                $user_id,
                                 $srs_rate,
                                 $srs_sequence,
                                 $srs_days,
